Removing controller, views and tests for schemes. Showing login/logout info for users
