+++ app/assets/stylesheets/base.css.scss
@@ -1,0 +1,24 @@
+html, body {
+  margin: 0;
+  padding: 0;
+}
+
+#topBar {
+  background-color: black;
+  color: #aaaaaa;
+  text-align: right;
+  padding: 0;
+  margin: 0;
+  .in
+  {
+    color: yellow;
+  }
+  a 
+  {
+    color: lime;
+  }
+}
+
+#content { 
+  padding: 1em;
+}
