--- app/assets/stylesheets/schemes.css.scss
@@ -1,18 +1,0 @@
-// Place all the styles related to the Schemes controller here.
-// They will automatically be included in application.css.
-// You can use Sass (SCSS) here: http://sass-lang.com/
-table#schemes-list {
-	thead {
-		background-color: black;
-		color: white;
-		td {padding: 10px; margin: 0px; font-size: 1.1em; opacity: 0.8; text-align: center}
-	
-	}
-	th {
-		text-align: left;
-	}
-	td {
-		padding: 5px;
-	}
-	border: 1px solid black;
-}
