+++ app/helpers/services_helper.rb
@@ -1,2 +1,10 @@
 module ServicesHelper
+	def show_amount(amount)
+		return number_to_currency(amount) if amount > 0
+	  		"free"
+	end
+
+	def show_frequency(frequency)
+		return frequency.first.capitalize
+	end	
 end
