+++ app/views/layouts/application.html.erb
@@ -8,7 +8,17 @@
 </head>
 <body>
 
+<div id="topBar">
+<% if user_signed_in? %>
+  <span class="in"><%= current_user.username %> | <%= link_to 'Logout', destroy_user_session_path, method: :delete %></span>
+<% else %>
+  <%= link_to 'Login', new_user_session_path %> 
+<% end %>
+</div>
+
+<div id="content">
 <%= yield %>
+</div>
 
 </body>
 </html>
