--- app/views/schemes/_form.html.erb
@@ -1,33 +1,0 @@
-<%= form_for(@scheme) do |f| %>
-  <% if @scheme.errors.any? %>
-    <div id="error_explanation">
-      <h2><%= pluralize(@scheme.errors.count, "error") %> prohibited this scheme from being saved:</h2>
-
-      <ul>
-      <% @scheme.errors.full_messages.each do |msg| %>
-        <li><%= msg %></li>
-      <% end %>
-      </ul>
-    </div>
-  <% end %>
-
-  <div class="field">
-    <%= f.label :name %><br />
-    <%= f.text_field :name %>
-  </div>
-  <div class="field">
-    <%= f.label :amount %><br />
-    <%= f.text_field :amount %>
-  </div>
-  <div class="field">
-    <%= f.label :currency %><br />
-    <%= f.text_field :currency %>
-  </div>
-  <div class="field">
-    <%= f.label :frequency %><br />
-    <%= f.text_field :frequency %>
-  </div>
-  <div class="actions">
-    <%= f.submit %>
-  </div>
-<% end %>
