--- app/views/schemes/edit.html.erb
@@ -1,6 +1,0 @@
-<h1>Editing scheme</h1>
-
-<%= render 'form' %>
-
-<%= link_to 'Show', @scheme %> |
-<%= link_to 'Back', :back %>
