--- app/views/schemes/new.html.erb
@@ -1,5 +1,0 @@
-<h1>New scheme</h1>
-
-<%= render 'form' %>
-
-<%= link_to 'Back', schemes_path %>
