--- app/views/schemes/show.html.erb
@@ -1,25 +1,0 @@
-<p id="notice"><%= notice %></p>
-
-<p>
-  <b>Name:</b>
-  <%= @scheme.name %>
-</p>
-
-<p>
-  <b>Amount:</b>
-  <%= @scheme.amount %>
-</p>
-
-<p>
-  <b>Currency:</b>
-  <%= @scheme.currency %>
-</p>
-
-<p>
-  <b>Frequency:</b>
-  <%= @scheme.frequency %>
-</p>
-
-
-<%= link_to 'Edit', edit_scheme_path(@scheme) %> |
-<%= link_to 'Back', :back %>
