+++ app/views/users/_form.html.erb
@@ -17,8 +17,13 @@
   </div>
   <div class="field">
     <%= f.label :email %><br />
-    <%= f.text_field :email %>
+    <%= f.email_field :email %>
   </div>
+  <div class="field">
+    <%= f.label :password %><br />
+    <%= f.password_field :password %>
+  </div>
+  
   <div class="actions">
     <%= f.submit %>
   </div>
