+++ app/views/users/_user.html.erb
@@ -1,0 +1,12 @@
+<p>
+  <b>Username:</b>
+  <%= @user.username %>
+</p>
+<p>
+  <b>Email:</b>
+  <%= @user.email %>
+</p>
+<p>
+  <b>Subscriptions:</b>
+  <%= @user.subscriptions.count %> (<%= link_to 'See', user_subscriptions_path(@user) %>)
+</p>
