+++ app/views/users/dashboard.html.erb
@@ -1,0 +1,5 @@
+<h1>User dashboard</h1>
+
+<%= render @user %>
+
+<p><%= link_to 'Services', services_path %></p>
