--- spec/controllers/schemes_controller_spec.rb
@@ -1,161 +1,0 @@
-require 'spec_helper'
-
-# This spec was generated by rspec-rails when you ran the scaffold generator.
-# It demonstrates how one might use RSpec to specify the controller code that
-# was generated by Rails when you ran the scaffold generator.
-#
-# It assumes that the implementation code is generated by the rails scaffold
-# generator.  If you are using any extension libraries to generate different
-# controller code, this generated spec may or may not pass.
-#
-# It only uses APIs available in rails and/or rspec-rails.  There are a number
-# of tools you can use to make these specs even more expressive, but we're
-# sticking to rails and rspec-rails APIs to keep things simple and stable.
-#
-# Compared to earlier versions of this generator, there is very limited use of
-# stubs and message expectations in this spec.  Stubs are only used when there
-# is no simpler way to get a handle on the object needed for the example.
-# Message expectations are only used when there is no simpler way to specify
-# that an instance is receiving a specific message.
-
-describe SchemesController do
-
-   pending "Add or remove once embedded is sorted."
-=begin Commented until ready 
-
-  # This should return the minimal set of attributes required to create a valid
-  # Scheme. As you add validations to Scheme, be sure to
-  # update the return value of this method accordingly.
-  def valid_attributes
-    {}
-  end
-
-  describe "GET index" do
-    it "assigns all schemes as @schemes" do
-      scheme = Scheme.create! valid_attributes
-      get :index
-      assigns(:schemes).should eq([scheme])
-    end
-  end
-
-  describe "GET show" do
-    it "assigns the requested scheme as @scheme" do
-      scheme = Scheme.create! valid_attributes
-      get :show, :id => scheme.id
-      assigns(:scheme).should eq(scheme)
-    end
-  end
-
-  describe "GET new" do
-    it "assigns a new scheme as @scheme" do
-      get :new
-      assigns(:scheme).should be_a_new(Scheme)
-    end
-  end
-
-  describe "GET edit" do
-    it "assigns the requested scheme as @scheme" do
-      scheme = Scheme.create! valid_attributes
-      get :edit, :id => scheme.id
-      assigns(:scheme).should eq(scheme)
-    end
-  end
-
-  describe "POST create" do
-    describe "with valid params" do
-      it "creates a new Scheme" do
-        expect {
-          post :create, :scheme => valid_attributes
-        }.to change(Scheme, :count).by(1)
-      end
-
-      it "assigns a newly created scheme as @scheme" do
-        post :create, :scheme => valid_attributes
-        assigns(:scheme).should be_a(Scheme)
-        assigns(:scheme).should be_persisted
-      end
-
-      it "redirects to the created scheme" do
-        post :create, :scheme => valid_attributes
-        response.should redirect_to(Scheme.last)
-      end
-    end
-
-    describe "with invalid params" do
-      it "assigns a newly created but unsaved scheme as @scheme" do
-        # Trigger the behavior that occurs when invalid params are submitted
-        Scheme.any_instance.stubs(:save).returns(false)
-        post :create, :scheme => {}
-        assigns(:scheme).should be_a_new(Scheme)
-      end
-
-      it "re-renders the 'new' template" do
-        # Trigger the behavior that occurs when invalid params are submitted
-        Scheme.any_instance.stubs(:save).returns(false)
-        post :create, :scheme => {}
-        response.should render_template("new")
-      end
-    end
-  end
-
-  describe "PUT update" do
-    describe "with valid params" do
-      it "updates the requested scheme" do
-        scheme = Scheme.create! valid_attributes
-        # Assuming there are no other schemes in the database, this
-        # specifies that the Scheme created on the previous line
-        # receives the :update_attributes message with whatever params are
-        # submitted in the request.
-        Scheme.any_instance.expects(:update_attributes).with({'these' => 'params'})
-        put :update, :id => scheme.id, :scheme => {'these' => 'params'}
-      end
-
-      it "assigns the requested scheme as @scheme" do
-        scheme = Scheme.create! valid_attributes
-        put :update, :id => scheme.id, :scheme => valid_attributes
-        assigns(:scheme).should eq(scheme)
-      end
-
-      it "redirects to the scheme" do
-        scheme = Scheme.create! valid_attributes
-        put :update, :id => scheme.id, :scheme => valid_attributes
-        response.should redirect_to(scheme)
-      end
-    end
-
-    describe "with invalid params" do
-      it "assigns the scheme as @scheme" do
-        scheme = Scheme.create! valid_attributes
-        # Trigger the behavior that occurs when invalid params are submitted
-        Scheme.any_instance.stubs(:save).returns(false)
-        put :update, :id => scheme.id, :scheme => {}
-        assigns(:scheme).should eq(scheme)
-      end
-
-      it "re-renders the 'edit' template" do
-        scheme = Scheme.create! valid_attributes
-        # Trigger the behavior that occurs when invalid params are submitted
-        Scheme.any_instance.stubs(:save).returns(false)
-        put :update, :id => scheme.id, :scheme => {}
-        response.should render_template("edit")
-      end
-    end
-  end
-
-  describe "DELETE destroy" do
-    it "destroys the requested scheme" do
-      scheme = Scheme.create! valid_attributes
-      expect {
-        delete :destroy, :id => scheme.id
-      }.to change(Scheme, :count).by(-1)
-    end
-
-    it "redirects to the schemes list" do
-      scheme = Scheme.create! valid_attributes
-      delete :destroy, :id => scheme.id
-      response.should redirect_to(schemes_url)
-    end
-  end
-=end
-
-end
