--- spec/factories/schemes.rb
@@ -1,10 +1,0 @@
-# Read about factories at http://github.com/thoughtbot/factory_girl
-
-FactoryGirl.define do
-  factory :scheme do
-    name ""
-    amount ""
-    currency ""
-    frequency ""
-  end
-end
