+++ spec/factories/users.rb
@@ -1,3 +1,5 @@
 Factory.define :user do |user|
   user.username "username"
+  user.email "something@something.xxx"
+  user.password "123123"
 end
