--- spec/helpers/schemes_helper_spec.rb
@@ -1,15 +1,0 @@
-require 'spec_helper'
-
-# Specs in this file have access to a helper object that includes
-# the SchemesHelper. For example:
-#
-# describe SchemesHelper do
-#   describe "string concat" do
-#     it "concats two strings with spaces" do
-#       helper.concat_strings("this","that").should == "this that"
-#     end
-#   end
-# end
-describe SchemesHelper do
-  #pending "add some examples to (or delete) #{__FILE__}"
-end
