--- spec/requests/schemes_spec.rb
@@ -1,12 +1,0 @@
-require 'spec_helper'
-
-describe "Schemes" do
-  describe "GET /schemes" do
-    pending "Check once embedded is done."
-    #it "works! (now write some real specs)" do
-      # Run the generator again with the --webrat flag if you want to use webrat methods/matchers
-      #get schemes_path
-      #response.status.should be(200)
-    #end
-  end
-end
