--- spec/views/schemes/new.html.erb_spec.rb
@@ -1,23 +1,0 @@
-require 'spec_helper'
-
-describe "schemes/new.html.erb" do
-  before(:each) do
-    assign(:scheme, stub_model(Scheme,
-      :name => "",
-      :amount => "",
-      :currency => "",
-      :frequency => ""
-    ).as_new_record)
-  end
-
-  it "renders new scheme form" do
-    render
-
-    assert_select "form", :action => schemes_path, :method => "post" do
-      assert_select "input#scheme_name", :name => "scheme[name]"
-      assert_select "input#scheme_amount", :name => "scheme[amount]"
-      assert_select "input#scheme_currency", :name => "scheme[currency]"
-      assert_select "input#scheme_frequency", :name => "scheme[frequency]"
-    end
-  end
-end
